network changes no hidden files
